requirements.txt
jupyter
jupyter_contrib_nbextensions
notebook==6.4.12
traitlets==5.9.0
numpy
pandas
requests
tensorflow-cpu==2.11.1
pillow
tensorflow-serving-api==2.7.0
flask
matplotlib
scikit-learn
awscli
taipy